end of eggvangelion
egg throwing basically works. if it is the only order remaining at the time it will go into 'auto yeet' where it repeats the throw as long as they are holding an egg and have energy. this allows chains to be set up easily. essential for solo and 2 player games, but avoids some ui fighting

work towards nanite bomb implementation, currently not working